refactor: restructure incoming call handlers packages
